fix(itools): optimize nvim UX with previous extra tools included.
This commit serves to fix a minor issue that was occurring where a single one of
the files that is used in the overall nvim configuration would completely hang
and cause a freeze of the terminal window.
